examples/fedppo/pyproject.toml
[project]
name = "fedppo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "torch >=2.6.0",
    "torchrl[gym_continuous] >=0.7.0",
    "florl",
    "pydantic>=2.10.6",
    "hydra-core>=1.3.2",
    "flwr[simulation]>=1.15.0",
    "ray>=2.42.0", # flwr doesn't install ray on Python 3.12
    "wandb>=0.19.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
